Error Logging, User Management, Localization
This commit is large and important enough for a new release, so have a look at the release description for more information.

+++ Virgo/commands/clear.js
@@ -1,34 +1,58 @@
 const { SlashCommandBuilder, PermissionFlagsBits } = require('discord.js');
 const config = require('../assets/config.js');
+const userUtils = require('../utils/user.js');
+const logger = require('../utils/logger.js');
 
 module.exports = {
     cooldown: config.cooldowns.A,
     data: new SlashCommandBuilder()
         .setName('clear')
+        .setNameLocalizations({
+            nl: "opruimen"
+        })
         .setDescription('Bulk delete messages inside the current channel.')
-        .setDefaultMemberPermissions(PermissionFlagsBits.Administrator)
+        .setDescriptionLocalizations({
+            nl: "Berichten in bulk verwijderen binnen het huidige kanaal."
+        })
+        .setDefaultMemberPermissions(PermissionFlagsBits.ManageGuild)
         .addIntegerOption(option => option
             .setName('amount')
+            .setNameLocalizations({
+                nl: "hoeveelheid"
+            })
             .setDescription('Amount of messages to delete.')
+            .setDescriptionLocalizations({
+                nl: "De hoeveelheid berichten die verwijderd moeten worden."
+            })
             .setRequired(true)
             .setMinValue(1)
             .setMaxValue(50)),
     async execute(interaction) {
-        const amount = interaction.options.getInteger('amount');
-        interaction.reply({
-            content: `Deleting ${amount} messages . . .`,
-            ephemeral: true
-        });
-
-        // Bulk Delete
-        setTimeout(() => {
-            interaction.deleteReply();
-            interaction.channel.bulkDelete(amount).catch(() => {
-                interaction.reply({
-                    content: "Atleast one of the messages you are trying to delete is older than \`14\` days. Discord is not allowing me to do that, so you will have to delete them manually (or lower your clear amount to potentially exclude the erroneous message).",
-                    ephemeral: true
-                });
+        try {
+            // Permission Validation
+            if (!(await userUtils.checkAdmin(interaction.user.id, interaction.guild))) return interaction.reply({
+                content: `You do not have the required permissions to perform this elevated command. Please try again later, or contact moderation to receive elevated permissions.`,
+                ephemeral: true
             });
-        }, 1000);
+
+            const amount = interaction.options.getInteger('amount');
+            interaction.reply({
+                content: `Deleting ${amount} messages . . .`,
+                ephemeral: true
+            });
+
+            // Bulk Delete
+            setTimeout(() => {
+                interaction.deleteReply();
+                interaction.channel.bulkDelete(amount).catch(() => {
+                    interaction.reply({
+                        content: "Atleast one of the messages you are trying to delete is older than \`14\` days. Discord is not allowing me to do that, so you will have to delete them manually (or lower your clear amount to potentially exclude the erroneous message).",
+                        ephemeral: true
+                    });
+                });
+            }, 1000);
+        } catch (error) {
+            logger.error(error);
+        }
     }
 };
